gulpのリファクタ
