+++ manifest.json
@@ -4,10 +4,10 @@
     "version": "2.2",
     "description": "ニコニコ動画で動画をザッピングできるプレイヤーが立ち上がるChrome拡張機能。動画リンクをクリックすると起動します。ランキングや検索結果は勿論、コミュニティなどニコニコ動画のサブドメイン下ならどこでも起動します。",
     "permissions": [
-      "http://*/*",
-      "https://*/*",
-      "storage",
-      "tabs"
+        "http://*/*",
+        "https://*/*",
+        "storage",
+        "tabs"
     ],
     "icons": {
         "16": "dist/img/icon/16.png",
@@ -35,7 +35,9 @@
         }
     ],
     "background": {
-        "scripts": ["dist/background.js"],
+        "scripts": [
+            "dist/background.js"
+        ],
         "persistent": false
     },
     "web_accessible_resources": [
