+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "dependencies": {},
   "devDependencies": {
+    "@babel/register": "^7.5.5",
     "@createjs/easeljs": "^2.0.0-beta.4",
     "babel-core": "^6.26.3",
     "babel-loader": "^7.1.5",
@@ -17,6 +18,7 @@
     "eslint-loader": "^2.1.1",
     "exports-loader": "^0.7.0",
     "expose-loader": "^0.7.5",
+    "fs": "0.0.1-security",
     "gulp": "^4.0.0",
     "gulp-autoprefixer": "^6.0.0",
     "gulp-cli": "^2.0.1",
