+++ src/js/content/map/release-history.js
@@ -9,6 +9,10 @@ export default [
             '非ログイン時で見れないところを塞いだ（視聴履歴・ニコレポ・マイリスト登録ボタンなど）',
             '動画情報パネルの文字色等々、一部見づらかった場所のコントラストを高くした',
         ],
+        develop: [
+            'リリース履歴を切り出してそこからmanifest.jsonを更新するようにした',
+            '.editorconfigを追加した',
+        ],
     },
     {
         version: '2.1',
