+++ src/manifest-template.js
@@ -1,13 +1,13 @@
-{
+export default {
     "manifest_version": 2,
     "name": "niico",
-    "version": "",
+    "version": "*****更新履歴を元に書き換えられる*****",
     "description": "ニコニコ動画で動画をザッピングできるプレイヤーが立ち上がるChrome拡張機能。動画リンクをクリックすると起動します。ランキングや検索結果は勿論、コミュニティなどニコニコ動画のサブドメイン下ならどこでも起動します。",
     "permissions": [
-      "http://*/*",
-      "https://*/*",
-      "storage",
-      "tabs"
+        "http://*/*",
+        "https://*/*",
+        "storage",
+        "tabs"
     ],
     "icons": {
         "16": "dist/img/icon/16.png",
